cleanups, clarifications, fixups for alignRequests

+++ metric_tank/dataprocessor.go
@@ -134,8 +134,8 @@ func consolidate(in []Point, aggNum uint32, consolidator consolidation.Consolida
 
 // returns how many points should be aggregated together so that you end up with as many points as possible,
 // but never more than maxPoints
-func aggEvery(numPoints, maxPoints uint32) int {
-	return int((numPoints + maxPoints - 1) / maxPoints)
+func aggEvery(numPoints, maxPoints uint32) uint32 {
+	return (numPoints + maxPoints - 1) / maxPoints
 }
 
 func getTarget(req Req) (points []Point, interval uint32, err error) {
